Refactor customer-related methods and properties, including adding support for fetching a single customer by remote ID and displaying error messages in the customer list view model.

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/network/rest/wpcom/wc/customer/dto/CustomerFromAnalyticsDTO.kt
@@ -34,7 +34,7 @@ data class CustomerFromAnalyticsDTO(
     @SerializedName("total_spend")
     val totalSpend: Double? = null,
     @SerializedName("user_id")
-    val userId: Int? = null,
+    val userId: Long? = null,
     @SerializedName("username")
     val username: String? = null
 )
